consolidate matrix sources
